Create objective-c category and swift extensions; also
rename the category methods to include a prefix.

In stead of having to subclass from FBSnapshotTestCase, it is now
possible to use categories directly on XCTestCase through macros in
Ojbective-C, and in Swift protocol oriented programming is used to allow
each test class to be extended by adhering to the
FBSnapshotCapableTestCase protocol.